+++ CHANGELOG.md
@@ -2,13 +2,13 @@
 
 ## [3.0.0] - 2024-01-17
 
-## Added
+### Added
 
 - Support for specifying registry+image in one parameter (`--to` and `--from` as opposed to `--toRegistry` + `--toImage` etc)
 - Support for cross mounting of base imagelayers: https://docker-docs.uclv.cu/registry/spec/api/#cross-repository-blob-mount . This should save space in registries and be more efficient.
 - Support for GitLab CI-tokens
 
-## Fixed
+### Fixed
 
 - Uses `application/content-stream` as Content-Type when POSTing layers to registry as Github++ doesn't like us posting the actual media type and the docker description also suggests `application/content-stream` (https://docker-docs.uclv.cu/registry/spec/api/#monolithic-upload)
 - Handle relative URLs when posting to /blobs/uploads/